+++ test/junit/abfab3d/grid/query/TestCanMoveMaterial.java
@@ -86,10 +86,10 @@ public class TestCanMoveMaterial extends BaseTestGrid {
         Grid grid = new ArrayGridByte(100,100,100,0.001, 0.001);
 
         // set the voxels of a square
-        setX(grid, 50, 40, Grid.EXTERIOR, (byte) 1, 40, 60);
-        setX(grid, 50, 60, Grid.EXTERIOR, (byte) 1, 40, 60);
-        setZ(grid, 40, 50, Grid.EXTERIOR, (byte) 1, 40, 60);
-        setZ(grid, 60, 50, Grid.EXTERIOR, (byte) 1, 40, 60);
+        setX(grid, 50, 40, Grid.EXTERIOR, 1, 40, 60);
+        setX(grid, 50, 60, Grid.EXTERIOR, 1, 40, 60);
+        setZ(grid, 40, 50, Grid.EXTERIOR, 1, 40, 60);
+        setZ(grid, 60, 50, Grid.EXTERIOR, 1, 40, 60);
 
         // set the voxels of a T shape with the bottom of the T intersecting the opening of the square
         setX(grid, 60, 50, Grid.EXTERIOR, matToMove, 30, 70);
@@ -133,10 +133,10 @@ public class TestCanMoveMaterial extends BaseTestGrid {
         Grid grid = new ArrayGridByte(100,100,100,0.001, 0.001);
 
         // set the voxels of a square
-        setX(grid, 50, 40, Grid.EXTERIOR, (byte) 1, 40, 60);
-        setX(grid, 50, 60, Grid.EXTERIOR, (byte) 1, 40, 60);
-        setZ(grid, 40, 50, Grid.EXTERIOR, (byte) 1, 40, 60);
-        setZ(grid, 60, 50, Grid.EXTERIOR, (byte) 1, 40, 60);
+        setX(grid, 50, 40, Grid.EXTERIOR, 1, 40, 60);
+        setX(grid, 50, 60, Grid.EXTERIOR, 1, 40, 60);
+        setZ(grid, 40, 50, Grid.EXTERIOR, 1, 40, 60);
+        setZ(grid, 60, 50, Grid.EXTERIOR, 1, 40, 60);
 
         // set the voxels of an I shape with the vertical part intersecting the opening of the square
         setX(grid, 60, 50, Grid.EXTERIOR, matToMove, 30, 70);
@@ -175,9 +175,9 @@ public class TestCanMoveMaterial extends BaseTestGrid {
 
     /**
      * Test movement in all 26 paths
-     * 
+     *
      * 1) Set the center voxel to a material. This is the material to move.
-     * 2) Set the all the grid edge voxels to a different material. This 
+     * 2) Set the all the grid edge voxels to a different material. This
      *    should block the center material from escaping.
      * 3) For each path, reset the corresponding grid edge voxel back to
      *    outside and material 0, and then test movement in that path.
@@ -191,7 +191,7 @@ public class TestCanMoveMaterial extends BaseTestGrid {
 
         // set the material to move at the center of the grid
         grid.setData(center, center, center, Grid.EXTERIOR, material2);
-        
+
         // set the edge voxels of the grid to a different material
         setPlaneX(grid, 0, Grid.EXTERIOR, material1);
         setPlaneX(grid, size-1, Grid.EXTERIOR, material1);
@@ -207,65 +207,65 @@ public class TestCanMoveMaterial extends BaseTestGrid {
 
         // set the 26 paths
         int[][] paths = {
-        		{1,0,0},    {-1,0,0}, 
-        		{0,1,0},    {0,-1,0},
-        		{0,0,1},    {0,0,-1},
-        		{1,1,1},    {1,1,0},
-        		{1,1,-1},   {0,1,-1},
-        		{-1,1,-1},  {-1,1,0},
-        		{-1,1,1},   {0,1,1},
-        		{1,0,1},    {1,0,-1},
-        		{-1,0,-1},  {-1,0,1},
-        		{1,-1,1},   {1,-1,0},
-        		{1,-1,-1},  {0,-1,-1},
-        		{-1,-1,-1}, {-1,-1,0},
-        		{-1,-1,1},  {0,-1,1}
+                {1,0,0},    {-1,0,0},
+                {0,1,0},    {0,-1,0},
+                {0,0,1},    {0,0,-1},
+                {1,1,1},    {1,1,0},
+                {1,1,-1},   {0,1,-1},
+                {-1,1,-1},  {-1,1,0},
+                {-1,1,1},   {0,1,1},
+                {1,0,1},    {1,0,-1},
+                {-1,0,-1},  {-1,0,1},
+                {1,-1,1},   {1,-1,0},
+                {1,-1,-1},  {0,-1,-1},
+                {-1,-1,-1}, {-1,-1,0},
+                {-1,-1,1},  {0,-1,1}
         };
-        
+
         // the corresponding voxels that need to be reset to outside and
         // material 0 in order to allow the middle material to escape
         int[][] voxelsRemovedToAllowEscape = new int[paths.length][3];
-        
+
         for (int i=0; i<paths.length; i++) {
-        	for (int j=0; j<3; j++) {
-        		if (paths[i][j] > 0) {
-        			voxelsRemovedToAllowEscape[i][j] = endIndex;
-        		} else if (paths[i][j] < 0) {
-        			voxelsRemovedToAllowEscape[i][j] = 0;
-        		} else {
-        			voxelsRemovedToAllowEscape[i][j] = center;
-        		}
-        	}
+            for (int j=0; j<3; j++) {
+                if (paths[i][j] > 0) {
+                    voxelsRemovedToAllowEscape[i][j] = endIndex;
+                } else if (paths[i][j] < 0) {
+                    voxelsRemovedToAllowEscape[i][j] = 0;
+                } else {
+                    voxelsRemovedToAllowEscape[i][j] = center;
+                }
+            }
 //System.out.println("voxel: " + java.util.Arrays.toString(voxelsRemovedToAllowEscape[i]));
         }
-        
+
         boolean escaped;
-        
+
         // assert that movement is initially false in all paths
         for (int j=0; j<paths.length; j++) {
             escaped = canMove(grid, paths[j], material2);
             assertEquals(
-            		java.util.Arrays.toString(paths[j]) + " move is not false",
-            		false, 
-            		escaped);
+                    java.util.Arrays.toString(paths[j]) + " move is not false",
+                    false,
+                    escaped);
         }
 
-        
+
         // set the voxels to outside and material 0 in order to
         // allow escape in the corresponding path
         for (int j=0; j<paths.length; j++) {
-        	grid.setData(voxelsRemovedToAllowEscape[j][0], voxelsRemovedToAllowEscape[j][1], voxelsRemovedToAllowEscape[j][2], Grid.OUTSIDE, 0);
+            grid.setData(voxelsRemovedToAllowEscape[j][0], voxelsRemovedToAllowEscape[j][1], voxelsRemovedToAllowEscape[j][2], Grid.OUTSIDE, 0);
 
             escaped = canMove(grid, paths[j], material2);
-            
+
             assertEquals(
-            		java.util.Arrays.toString(paths[j]) + " move is not true",
-            		true, 
-            		escaped);
+                    java.util.Arrays.toString(paths[j]) + " move is not true",
+                    true,
+                    escaped);
         }
 
     }
-    
+
     public void testIgnoredVoxels() {
         int size = 12;
 
@@ -274,14 +274,14 @@ public class TestCanMoveMaterial extends BaseTestGrid {
         // Add Object 1
         int mat1_count = 5;
 
-        grid.setData(0,0,0, Grid.EXTERIOR, (byte) 1);
-        grid.setData(0,0,1, Grid.INTERIOR, (byte) 1);
-        grid.setData(0,0,2, Grid.INTERIOR, (byte) 1);
-        grid.setData(0,0,3, Grid.INTERIOR, (byte) 1);
-        grid.setData(0,0,4, Grid.EXTERIOR, (byte) 1);
+        grid.setData(0,0,0, Grid.EXTERIOR, 1);
+        grid.setData(0,0,1, Grid.INTERIOR, 1);
+        grid.setData(0,0,2, Grid.INTERIOR, 1);
+        grid.setData(0,0,3, Grid.INTERIOR, 1);
+        grid.setData(0,0,4, Grid.EXTERIOR, 1);
 
         StraightPath path = new StraightPath(new int[] {0,0,-1});
-        CanMoveMaterial query = new CanMoveMaterial((byte) 1, path);
+        CanMoveMaterial query = new CanMoveMaterial(1, path);
         boolean escaped = query.execute(grid);
 
 //        assertTrue("X Axis move", escaped == true);
@@ -300,16 +300,16 @@ public class TestCanMoveMaterial extends BaseTestGrid {
 //          grid.setData(i,0,0, Grid.INTERIOR, (byte) 1);
 //        }
 
-        setX(grid, 0, 0, Grid.INTERIOR, (byte) 1, startIndex+1, endIndex-1);
-        setX(grid, 0, 0, Grid.OUTSIDE, (byte) 1, 101, 109);
+        setX(grid, 0, 0, Grid.INTERIOR, 1, startIndex+1, endIndex-1);
+        setX(grid, 0, 0, Grid.OUTSIDE, 1, 101, 109);
 
-        grid.setData(startIndex,0,0, Grid.EXTERIOR, (byte) 1);
-        grid.setData(endIndex,0,0, Grid.EXTERIOR, (byte) 1);
-        grid.setData(100,0,0, Grid.EXTERIOR, (byte) 1);
-        grid.setData(110,0,0, Grid.EXTERIOR, (byte) 1);
+        grid.setData(startIndex,0,0, Grid.EXTERIOR, 1);
+        grid.setData(endIndex,0,0, Grid.EXTERIOR, 1);
+        grid.setData(100,0,0, Grid.EXTERIOR, 1);
+        grid.setData(110,0,0, Grid.EXTERIOR, 1);
 
         StraightPath path = new StraightPath(new int[] {-1,0,0});
-        CanMoveMaterial query = new CanMoveMaterial((byte) 1, path);
+        CanMoveMaterial query = new CanMoveMaterial(1, path);
 
         long stime = System.nanoTime();
         query.execute(grid);
@@ -340,7 +340,7 @@ public class TestCanMoveMaterial extends BaseTestGrid {
         long stime = System.nanoTime();
 
         for (int j=0; j<paths.length; j++) {
-            queryOnePath = new CanMoveMaterial((byte) 1, paths[j]);
+            queryOnePath = new CanMoveMaterial(1, paths[j]);
             queryOnePath.execute(grid);
         }
 
@@ -352,7 +352,7 @@ public class TestCanMoveMaterial extends BaseTestGrid {
         //------------------------------------------------------------
         stime = System.nanoTime();
 
-        CanMoveMaterialAllPaths queryAllPaths = new CanMoveMaterialAllPaths((byte) 1, paths);
+        CanMoveMaterialAllPaths queryAllPaths = new CanMoveMaterialAllPaths(1, paths);
         queryAllPaths.execute(grid);
 
         totalTime1 = System.nanoTime() - stime;
@@ -374,19 +374,19 @@ public class TestCanMoveMaterial extends BaseTestGrid {
 
         Grid grid = new ArrayGridByte(size,size,size,0.001, 0.001);
 
-        setX(grid, yIndex, 0, Grid.INTERIOR, (byte) 1, startIndex+1, endIndex-1);
-        setX(grid, yIndex, 0, Grid.OUTSIDE, (byte) 0, 11, 12);
+        setX(grid, yIndex, 0, Grid.INTERIOR, 1, startIndex+1, endIndex-1);
+        setX(grid, yIndex, 0, Grid.OUTSIDE, 0, 11, 12);
 
-        grid.setData(startIndex,yIndex,0, Grid.EXTERIOR, (byte) 1);
-        grid.setData(endIndex,yIndex,0, Grid.EXTERIOR, (byte) 1);
-        grid.setData(10,yIndex,0, Grid.EXTERIOR, (byte) 1);
-        grid.setData(13,yIndex,0, Grid.EXTERIOR, (byte) 1);
+        grid.setData(startIndex,yIndex,0, Grid.EXTERIOR, 1);
+        grid.setData(endIndex,yIndex,0, Grid.EXTERIOR, 1);
+        grid.setData(10,yIndex,0, Grid.EXTERIOR, 1);
+        grid.setData(13,yIndex,0, Grid.EXTERIOR, 1);
 
         // Set different material
 //        grid.setData(5,yIndex,0, Grid.EXTERIOR, (byte) 2);
-        grid.setData(18,yIndex,0, Grid.EXTERIOR, (byte) 2);
-        grid.setData(7,5,0, Grid.EXTERIOR, (byte) 2);
-        grid.setData(10,15,0, Grid.EXTERIOR, (byte) 2);
+        grid.setData(18,yIndex,0, Grid.EXTERIOR, 2);
+        grid.setData(7,5,0, Grid.EXTERIOR, 2);
+        grid.setData(10,15,0, Grid.EXTERIOR, 2);
 
         return grid;
     }
@@ -399,19 +399,19 @@ public class TestCanMoveMaterial extends BaseTestGrid {
 
         Grid grid = new ArrayGridByte(size,size,size,0.001, 0.001);
 
-        setX(grid, yIndex, 0, Grid.INTERIOR, (byte) 1, startIndex+1, endIndex-1);
-        setX(grid, yIndex, 0, Grid.OUTSIDE, (byte) 0, 201, 249);
+        setX(grid, yIndex, 0, Grid.INTERIOR, 1, startIndex+1, endIndex-1);
+        setX(grid, yIndex, 0, Grid.OUTSIDE, 0, 201, 249);
 
-        grid.setData(startIndex,yIndex,0, Grid.EXTERIOR, (byte) 1);
-        grid.setData(endIndex,yIndex,0, Grid.EXTERIOR, (byte) 1);
-        grid.setData(200,yIndex,0, Grid.EXTERIOR, (byte) 1);
-        grid.setData(250,yIndex,0, Grid.EXTERIOR, (byte) 1);
+        grid.setData(startIndex,yIndex,0, Grid.EXTERIOR, 1);
+        grid.setData(endIndex,yIndex,0, Grid.EXTERIOR, 1);
+        grid.setData(200,yIndex,0, Grid.EXTERIOR, 1);
+        grid.setData(250,yIndex,0, Grid.EXTERIOR, 1);
 
         // Set different material
 //        grid.setData(50,yIndex,0, Grid.EXTERIOR, (byte) 2);
-        grid.setData(400,yIndex,0, Grid.EXTERIOR, (byte) 2);
-        grid.setData(100,5,0, Grid.EXTERIOR, (byte) 2);
-        grid.setData(200,15,0, Grid.EXTERIOR, (byte) 2);
+        grid.setData(400,yIndex,0, Grid.EXTERIOR, 2);
+        grid.setData(100,5,0, Grid.EXTERIOR, 2);
+        grid.setData(200,15,0, Grid.EXTERIOR, 2);
 
         return grid;
     }
